Rename package 'org.triplea.map.reader.generic.xml' -> 'org.triplea.generic.xml.reader'
