+++ map-data/src/main/java/org/triplea/map/data/elements/AttachmentList.java
@@ -2,8 +2,8 @@ package org.triplea.map.data.elements;
 
 import java.util.List;
 import lombok.Getter;
-import org.triplea.map.reader.generic.xml.Attribute;
-import org.triplea.map.reader.generic.xml.TagList;
+import org.triplea.generic.xml.reader.Attribute;
+import org.triplea.generic.xml.reader.TagList;
 
 @Getter
 public class AttachmentList {
