+++ map-data/src/main/java/org/triplea/map/data/elements/Triplea.java
@@ -1,7 +1,7 @@
 package org.triplea.map.data.elements;
 
 import lombok.Getter;
-import org.triplea.map.reader.generic.xml.Attribute;
+import org.triplea.generic.xml.reader.Attribute;
 
 @Getter
 public class Triplea {
