+++ xml-reader/src/main/java/org/triplea/generic/xml/reader/Attribute.java
@@ -1,7 +1,9 @@
-package org.triplea.map.reader.generic.xml;
+package org.triplea.generic.xml.reader;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Tag {}
+public @interface Attribute {
+  String defaultValue() default "";
+}
