+++ xml-reader/src/main/java/org/triplea/generic/xml/reader/Tag.java
@@ -1,9 +1,7 @@
-package org.triplea.map.reader.generic.xml;
+package org.triplea.generic.xml.reader;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 @Retention(RetentionPolicy.RUNTIME)
-public @interface TagList {
-  Class<?> value();
-}
+public @interface Tag {}
