+++ xml-reader/src/main/java/org/triplea/generic/xml/reader/TagList.java
@@ -1,9 +1,9 @@
-package org.triplea.map.reader.generic.xml;
+package org.triplea.generic.xml.reader;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Attribute {
-  String defaultValue() default "";
+public @interface TagList {
+  Class<?> value();
 }
